client/requirements.txt
ultralytics
opencv-python
websocket-client
requests
numpy
